BMP-157 remove clients directory, consolidate all external calls under providers, added common entry parameters to jambase and ticketmaster
